:recycle: [Refactor]: config 폴더로 설정파일 이동
