+++ src/main/java/PhishingUniv/Phinocchio/config/FCMConfig.java
@@ -1,4 +1,4 @@
-package PhishingUniv.Phinocchio.domain.Twilio.config;
+package PhishingUniv.Phinocchio.config;
 
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.firebase.FirebaseApp;
